% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_xptr}
\alias{create_xptr}
\title{Create external pointer to a C++ function for \code{f}}
\usage{
create_xptr(fstr)
}
\arguments{
\item{fstr}{A string indicating the C++ function required.}
}
\description{
Create external pointer to a C++ function for \code{f}
}
