% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{households}
\alias{households}
\title{Households data}
\format{
A tibble (see \link[tibble:tibble-package]{tibble::tibble-package}) with columns:
\describe{
\item{id_hh}{Household identifier.}
\item{ecv_year}{The year the survey was conducted. Data refer to
the previous year.}
\item{region}{Autonomous Community indicator. Values:
\tabular{ll}{
AND \tab Andalucía \cr
ARA \tab Aragón \cr
AST \tab Asturias \cr
BAL \tab Baleares \cr
CNR \tab Canarias \cr
CNT \tab Cantabria \cr
CYL \tab Castilla y León \cr
CLM \tab Castilla La Mancha\cr
CAT \tab Cataluña \cr
VAL \tab Valencia \cr
EXT \tab Extremadura \cr
GAL \tab Galicia \cr
MAD \tab Madrid \cr
MUR \tab Murcia \cr
NAV \tab Navarra \cr
PVA \tab País Vasco \cr
RIO \tab La Rioja \cr
CEU \tab Ceuta \cr
MEL \tab Melilla \cr
}
}
\item{cluster}{Primary sampling unit indicator.}
\item{weight_hh}{Household population weight.}
\item{urb}{Degree of urbanisation. Factor levels: 'Low', 'Med', and 'High'.}
\item{size_hh}{Number of persons per household}
\item{type_hh}{Household type. Factor levels:
\tabular{ll}{
HH_01 \tab One member: a man aged less than 30. \cr
HH_02 \tab One member: a man aged between 30 and 64. \cr
HH_03 \tab One member: a man aged 65 or more. \cr
HH_04 \tab One member: a woman aged less than 30. \cr
HH_05 \tab One member: a woman aged between 30 and 64. \cr
HH_06 \tab One member: a woman aged 65 or more. \cr
HH_07 \tab Two adults without dependent children and at least
one person aged 65 or more. \cr
HH_08 \tab Two adults without dependent children and both of them
aged less than 65.  \cr
HH_09 \tab Other households without dependent children. \cr
HH_10 \tab One adult with at least one dependent child. \cr
HH_11 \tab Two adults with one dependent child. \cr
HH_12 \tab Two adults with two dependent children. \cr
HH_13 \tab Two adults with three or more dependent children.  \cr
HH_14 \tab Other households with dependent children.\cr
}
}
\item{cunits_hh}{Consumption units: Modified OECD equivalence scale.}
\item{ydisp_hh}{Household disposable income.}
\item{pov_hh}{At risk of poverty indicator.}
\item{depriv_hh}{Severe material deprivation indicator.}
}
}
\source{
\href{https://ine.es/dyngs/INEbase/en/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=resultados&secc=1254736195153&idp=1254735976608#!tabs-1254736195153}{Microdata at INE, the Spanish Statistical Office}
}
\usage{
households
}
\description{
Household data from the ECV for the years 2005, 2011, and 2019.
}
\keyword{datasets}
