% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individuals}
\alias{individuals}
\title{Individuals data}
\format{
A tibble (see \link[tibble:tibble-package]{tibble::tibble-package}).
}
\source{
\href{https://ine.es/dyngs/INEbase/en/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=resultados&secc=1254736195153&idp=1254735976608#!tabs-1254736195153}{Microdata at INE, the Spanish Statistical Office}
}
\usage{
individuals
}
\description{
Personal data from the ECV for the years 2005, 2011, and 2019.
}
\keyword{datasets}
