% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itp-package.R
\docType{package}
\name{itp-package}
\alias{itp}
\alias{itp-package}
\title{itp: Intergenerational Transmission of Poverty}
\description{
Data and code to study poverty and the transmission of poverty 
    across generations in Spain. Data is taken from the Spanish Encuesta de 
    Condiciones de Vida (ECV) collected by the Spanish Statistical Office, INE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcpernias/itp}
  \item Report bugs at \url{https://github.com/jcpernias/itp/issues}
}

}
\author{
\strong{Maintainer}: José Pernías \email{jcpernias@gmail.com}

}
\keyword{internal}
