% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pov_trans}
\alias{pov_trans}
\title{Intergenerational transmission of poverty data}
\format{
A tibble (see \link[tibble:tibble-package]{tibble::tibble-package}) with columns:
\describe{
\item{id_p}{Person identifier.}
\item{ecv_year}{The year the survey was conducted. Data refer to
the previous year.}
\item{age}{Age, in years.}
\item{mabsent}{}
\item{fabsent}{}
\item{adults}{}
\item{children}{}
\item{siblings}{}
\item{nworking}{}
\item{fcountry}{}
\item{fnation}{}
\item{mcountry}{}
\item{mnation}{}
\item{feduc}{}
\item{meduc}{}
\item{fworker}{}
\item{fself}{}
\item{funempl}{}
\item{finactive}{}
\item{foccup}{}
\item{mworker}{}
\item{mself}{}
\item{munempl}{}
\item{minactive}{}
\item{moccup}{}
\item{well_being}{}
\item{fin_hardship}{}
}
}
\source{
\href{https://ine.es/dyngs/INEbase/en/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=resultados&secc=1254736195153&idp=1254735976608#!tabs-1254736195153}{Microdata at INE, the Spanish Statistical Office}
}
\usage{
pov_trans
}
\description{
Data from the ECV for the years 2005, 2011, and 2019.
}
\keyword{datasets}
