% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{households}
\alias{households}
\title{Households data}
\format{
A \href{tibble::tibble-package}{tibble}.
\describe{
\item{id_hh}{Household identifier.}
\item{ecv_year}{The year the survey was conducted. Variables refer to
the previous year.}
}
}
\source{
\href{https://ine.es/dyngs/INEbase/en/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=resultados&secc=1254736195153&idp=1254735976608#!tabs-1254736195153}{Microdata at INE, the Spanish Statistical Office}
}
\usage{
households
}
\description{
Household data from the ECV for the years 2005, 2011, and 2019.
}
\keyword{datasets}
