% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individuals}
\alias{individuals}
\title{Individuals data}
\format{
A tibble (see \link[tibble:tibble-package]{tibble::tibble-package}) with columns:
\describe{
\item{id_p}{Person identifier.}
\item{id_hh}{Household identifier.}
\item{ecv_year}{The year the survey was conducted. Data refer to
the previous year.}
\item{birth_year}{The year of birth.}
\item{age}{Age, in years.}
\item{sex}{Factor with levels \code{Female} and \code{Male}.}
\item{absent}{Is the individual temporarily absent from the household?}
\item{lowjob}{Does the individual live in a household with low work intensity?}
\item{eu2020}{Europe 2020 indicator on poverty and social exclusion.}
\item{weight_p16}{Weight for individuals over 16 years old.}
\item{country_birth}{Country of birth. Factor with levels:
\tabular{ll}{
Spain	\tab Spain. \cr
EU \tab European Union (EU 28). \cr
Other	\tab Rest of the world. \cr
}}
\item{nationality}{Nationality. Factor with levels: see \code{country_birth}.}
\item{educ_year}{Year in which the individual achieved the highest educational level.}
\item{educ}{Education level. Factor with levels:
\tabular{ll}{
No	\tab Less than primary education. \cr
Prim \tab Primary education. \cr
LowerSec	\tab Lower secondary education. \cr
UpperSec	\tab Upper secondary education. \cr
Sup	\tab Tertiary education. \cr
}}
\item{work_age}{Age at which the individual started working.}
\item{work_years}{Years working.}
\item{full_time}{Working full-time.}
\item{part_time}{Working part-time.}
\item{self}{Self-employed?}
\item{worker}{Employee?}
\item{unempl}{Unemployed?}
\item{inactive}{Inactive?}
\item{has_worked}{Has the individual ever worked?}
\item{occup}{ISCO-08 occupation major group. Factor levels:
\tabular{ll}{
ISCO-1	\tab Managers. \cr
ISCO-2	\tab Professionals. \cr
ISCO-3	\tab Technicians and Associate Professionals. \cr
ISCO-4	\tab Clerical Support Workers. \cr
ISCO-5	\tab Services and Sales Workers. \cr
ISCO-6	\tab Skilled Agricultural, Forestry and Fishery Workers. \cr
ISCO-7	\tab Craft and Related Trades Workers. \cr
ISCO-8	\tab Plant and Machine Operators and Assemblers. \cr
ISCO-9	\tab Elementary Occupations. \cr
ISCO-0	\tab Armed Forces Occupations. \cr
}}
}
}
\source{
\href{https://ine.es/dyngs/INEbase/en/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=resultados&secc=1254736195153&idp=1254735976608#!tabs-1254736195153}{Microdata at INE, the Spanish Statistical Office}
}
\usage{
individuals
}
\description{
Personal data from the ECV for the years 2005, 2011, and 2019.
}
\keyword{datasets}
