% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itp-internal.R
\name{itp-internal}
\alias{itp-internal}
\alias{wiki}
\alias{lambert}
\alias{trig1}
\alias{trig2}
\alias{poly1}
\alias{poly2}
\alias{exponential}
\alias{logarithmic}
\alias{posynomial}
\alias{weierstrass}
\alias{polyfrac}
\alias{normalcdf}
\alias{normalpdf}
\alias{poly3}
\alias{exp_poly}
\alias{tan_poly}
\alias{circles}
\alias{stepfn}
\alias{geometric}
\alias{truncpoly}
\alias{staircase}
\alias{noisy_line}
\alias{warsaw}
\alias{sawtooth}
\alias{sawtooth3}
\title{Internal itp functions}
\usage{
wiki(x)

lambert(x)

trig1(x)

trig2(x)

poly1(x)

poly2(x)

exponential(x)

logarithmic(x)

posynomial(x)

weierstrass(x)

polyfrac(x)

normalcdf(x)

normalpdf(x)

poly3(x)

exp_poly(x)

tan_poly(x)

circles(x)

stepfn(x)

geometric(x)

truncpoly(x)

staircase(x)

noisy_line(x)

warsaw(x)

sawtooth(x)

sawtooth3(x)
}
\description{
Internal itp functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
