% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itp-internal.R
\name{itp-internal}
\alias{itp-internal}
\alias{itp_r}
\title{Internal itp functions}
\usage{
itp_r(f, ..., a, b, ya, yb, epsilon, k1, k2, for_rk, inc)
}
\description{
Internal itp functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
